Add "negative series handling" option to control stacking behaviour
Currently, each series is sampled to decide whether it is predominantly positive or negative. This is then used to group the series separately for stacking. This causes issues when dealing with series which fluctuate around 0, as sometimes they are grouped with positive series and sometimes they are not. For example, flows of electricity in and out of a country can be positive and negative, but when stacked with the electricity generation curves for the country, the stacking behaviour is inconsistent.

However, looking at previous GitHub issues, it seems that for other applications, this is the correct behaviour. I think one example was displaying network transmission, where Tx and Rx should never be stacked together.

Given that there doesn't seem to be a one-size-fits-all solution to whether to group these series, I figured it would be best to make it configurable so the user can choose. I've currently called it "negative series handling" but it feels like there might be a better name for it.